comentários

+++ src/EstruturaDados/CGrafo.cpp
@@ -49,7 +49,7 @@ void CGrafo::Write ( std::ostream& out ) const
    out.setf ( ios::left );
 
    // Tipo de grafo
-   out << setw ( 5 ) << static_cast<uint16_t> ( Tipo() ) ;
+   out << setw ( 5 ) << static_cast<uint8_t> ( Tipo() ) ;
 
    // Numero de objetos
    out << ' '<< setw ( 5 ) << objeto.size() << endl;
