+++ src/EstruturaDados/ObjetoGrafo/CObjetoGrafo_2VetoresConexoes.cpp
@@ -84,7 +84,7 @@ ostream& CObjetoGrafo_2VetoresConexoes::Write ( ostream& out ) const
 {
     out.setf ( ios::left );
     // Tipo de contorno
-    out << setw ( 5 ) << static_cast<uint16_t> ( Tipo() ) ;
+    out << setw ( 5 ) << static_cast<uint8_t> ( Tipo() ) ;
 
     // Rótulo de this
     out << ' ' << setw ( 5 ) << rotulo;
